% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{apiUrl}
\alias{apiUrl}
\title{Internal constant in my package}
\format{
An object of class \code{character} of length 1.
}
\usage{
apiUrl
}
\description{
This is an internal constant that should not be modified or accessed
directly outside of the package.
}
\keyword{internal}
